Refactored Searchbar into it's own component
