+++ App/pages/Homepage.js
@@ -4,32 +4,20 @@ import Topcomp from './components/homePage/Topcomp';
 import Listcomp from './components/homePage/Listcomp';
 import Blockcomp from './components/homePage/Blockcomp';
 import Mapcomp from './components/homePage/Mapcomp';
+import Searchbar from './components/homePage/Searchbar';
 
 import shops from './components/dummy-data';
 
-const searchFilter = (expression, collection) => {
-    // Given an expression/text and a collection of strings, for every element in collection
-    // check if (sub)string from collection matches expression
-
-    // eg, if collection.element = bulten, expression = bul ----> bul === bul // true
-
-    // Match shop names against expression (match all against empty string)
-    const partialMatch = ({name}) => {
-        const partialName = name.slice(0, expression.length);
-        const lowerCaseExpression = expression.toLowerCase();
-        const lowerCasePartialName = partialName.toLowerCase();
-        return '' === lowerCaseExpression || lowerCasePartialName === lowerCaseExpression;
-    };
-
-    // Return all partial matches in collection
-    return collection.filter(partialMatch);
-};
-
 export default class Homepage extends Component {
+
     state = {
         toggle: 'LV',
-        filteredShops: shops,
+        shops: shops,
     };
+
+    // Apply callback to shops, updating current shops
+    updateShops = (callback) => this.setState({shops: callback(shops)});
+
     _pressLV = () => this.setState({ toggle: 'LV' });
     _pressBV = () => this.setState({ toggle: 'BV' });
     _pressMV = () => this.setState({ toggle: 'MV' });
@@ -59,16 +47,13 @@ export default class Homepage extends Component {
                     />
                 </View>
 
-                <TextInput 
-                onChangeText = {(search) => {
-                    this.setState({filteredShops: searchFilter(search, shops)});
-                }}
-                placeholder = 'Placeholder'
+                <Searchbar 
+                    onChange={this.updateShops}
                 />
 
-                {this.state.toggle == 'LV' && <Listcomp shops={this.state.filteredShops}/>}
-                {this.state.toggle == 'BV' && <Blockcomp shops={this.state.filteredShops}/>}
-                {this.state.toggle == 'MV' && <Mapcomp shops={this.state.filteredShops}/>}
+                {this.state.toggle == 'LV' && <Listcomp shops={this.state.shops}/>}
+                {this.state.toggle == 'BV' && <Blockcomp shops={this.state.shops}/>}
+                {this.state.toggle == 'MV' && <Mapcomp shops={this.state.shops}/>}
             </View>
         );
     }
